Utilitybill-main/utilitybill_backend/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.9.0
Pillow==11.2.1
mysqlclient==2.2.7
